requirements.txt
black
dill
dvc[webdav]
hydra-colorlog==1.2.0
hydra-core==1.2.0
hydra-optuna-sweeper==1.2.0
mlflow
mlxtend
numpy
pandarallel
pandas
pathlib
pre-commit
pytest
python-dotenv
pytorch_lightning==1.7.1
pyrootutils
rich
rna
scipy
tensorboard
--pre --extra-index-url https://download.pytorch.org/whl/nightly/cu116
torch 
torchinfo
torchmetrics==0.9.3
torchvision
tqdm